Remove print_as_file and make Dimensions integral

+++ src/modules/geometry.cxx
@@ -16,10 +16,10 @@ struct Vector
 };
 
 export template<Numerical Component_type>
-using Vector_2d = Vector<Component_type, (uint8_t)2>;
+using Vector_2d = Vector<Component_type, 2>;
 
 export template<Numerical Component_type>
-using Vector_3d = Vector<Component_type, (uint8_t)3>;
+using Vector_3d = Vector<Component_type, 3>;
 
 export template<Numerical Coordinate_type, std::integral auto Dimensions>
 struct Point
@@ -29,10 +29,10 @@ struct Point
 };
 
 export template<Numerical Component_type>
-using Point_2d = Point<Component_type, (uint8_t)2>;
+using Point_2d = Point<Component_type, 2>;
 
 export template<Numerical Component_type>
-using Point_3d = Point<Component_type, (uint8_t)3>;
+using Point_3d = Point<Component_type, 3>;
 
 export template<Numerical Coordinate_type>
 struct Ray
@@ -41,6 +41,6 @@ struct Ray
     Vector_3d<Coordinate_type> velocity;
 
     Point_3d<Coordinate_type> at(Numerical auto t) const {
-        return Point_3d<Coordinate_type>{ .coordinates = origin.coordinates + velocity.components * t, .dimensions = (uint8_t)3}; 
+        return Point_3d<Coordinate_type>{ .coordinates = origin.coordinates + velocity.components * t, .dimensions = 3}; 
     }
 };
